refactor: extract stuff out of handleSubmit
